refactor: handled ecosystem dashboard and create ecosystem conditionally
