+++ src/components/Ecosystem/interfaces/index.ts
@@ -3,3 +3,16 @@ export interface IEcosystem {
     description: string
     logoUrl: string
 }
+
+export interface Ecosystem {
+    id: number
+    createDateTime: string
+    createdBy: number
+    lastChangedDateTime: string
+    lastChangedBy: number
+    name: string
+    description: string
+    logoUrl: string
+    website: string
+    roles: string[] 
+}
