--- src/components/Ecosystems/interfaces/index.ts
@@ -1,13 +1,0 @@
-export interface Ecosystem {
-    id: number
-    createDateTime: string
-    createdBy: number
-    lastChangedDateTime: string
-    lastChangedBy: number
-    name: string
-    description: string
-    logoUrl: string
-    website: string
-    roles: string[]
-    
-}
